requirements.txt
attrs==24.2.0
beautifulsoup4==4.12.3
build==1.2.2.post1
CacheControl==0.14.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cleo==2.1.0
crashtest==0.4.1
distlib==0.3.8
dulwich==0.21.7
exceptiongroup==1.2.2
fastjsonschema==2.20.0
filelock==3.16.1
h11==0.14.0
idna==3.10
importlib_metadata==8.5.0
installer==0.7.0
jaraco.classes==3.4.0
keyring==24.3.1
linkedin_api==2.2.0
linkedin_scraper==2.11.4
lxml==5.3.0
more-itertools==10.5.0
msgpack==1.1.0
outcome==1.3.0.post0
packaging==24.1
pexpect==4.9.0
pkginfo==1.11.1
platformdirs==4.3.6
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
ptyprocess==0.7.0
pycparser==2.22
pyproject_hooks==1.2.0
PySocks==1.7.1
python-dotenv==1.0.1
RapidFuzz==3.10.0
requests==2.32.3
requests-toolbelt==1.0.0
selenium==4.25.0
shellingham==1.5.4
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tomli==2.0.2
tomlkit==0.13.2
trio==0.26.2
trio-websocket==0.11.1
trove-classifiers==2024.9.12
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.26.6
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
xattr==1.1.0
zipp==3.20.2
pymongo==4.6.1
llama-index==0.11.2